Working on updating docs to handle new good filestructure
