+++ pyeo/apps/change_detection/simple_s2_change_detection.py
@@ -81,7 +81,7 @@ if __name__ == "__main__":
     # Atmospheric correction
     if args.do_preprocess or do_all:
         log.info("Applying sen2cor")
-        pyeo.preprocessing.atmospheric_correction(l1_image_path, l2_image_path, sen2cor_path, delete_unprocessed_image=False)
+        pyeo.sen2_funcs.atmospheric_correction(l1_image_path, l2_image_path, sen2cor_path, delete_unprocessed_image=False)
 
     # Merging / Aggregating layers into single image
     if args.do_merge or do_all:
