+++ pyeo/classification.py
@@ -1,3 +1,8 @@
+"""
+Contains every function to do with map classification. This includes model creation, map classification and processes
+for array manipulation into scikit-learn compatible forms.
+"""
+
 import glob
 import logging
 import os
