--- pyeo/tests/test_sen2Chain.py
@@ -1,3 +1,0 @@
-
-
-
